chore: move data_generators to domain module
