+++ ctmds/domain/commodity_price/models/price_generator/generic.py
@@ -1,13 +1,13 @@
 from datetime import datetime
 from typing import Callable
 
-from ctmds.data_generators.daily_price import daily_prices_with_timestamps
 from ctmds.domain.commodity_price.models.price import PriceCollection
 from ctmds.domain.commodity_price.models.price_generator.interface import (
     DailyPricesGeneratorInterface,
     PriceGenerator,
 )
 from ctmds.domain.constants import CountryCodes, Granularity
+from ctmds.domain.data_generators.daily_price import daily_prices_with_timestamps
 
 
 class GenericDailyPricesGenerator(DailyPricesGeneratorInterface):
